indian-farmer-assistance-app/ai-service/requirements.txt
# AI/ML Service Dependencies for Indian Farmer Assistance Application

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# ML/AI Libraries
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
sentence-transformers==2.2.2
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.2.0

# NLP and Language Processing
indic-nlp-library==0.71
aiodns==3.1.1

# Configuration
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
loguru==0.7.2
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
aiofiles==23.2.1
tenacity==8.2.3